fix abs for 64-bit
svn: r9492

+++ src/mzscheme/src/numarith.c
@@ -353,7 +353,7 @@ scheme_abs(int argc, Scheme_Object *argv[])
   o = argv[0];
 
   if (SCHEME_INTP(o)) {
-    int n = SCHEME_INT_VAL(o);
+    long n = SCHEME_INT_VAL(o);
     return scheme_make_integer_value(ABS(n));
   } 
   t = _SCHEME_TYPE(o);
